% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{blur}
\alias{blur}
\title{Blur}
\usage{
blur(input_path, output_path)
}
\arguments{
\item{input_path}{string, path for the input png file}

\item{output_path}{string, path for the output png file}
}
\value{
a png file at the specified output path
}
\description{
Blur
}
\examples{
#' blur("input.png", "blur.png")
}
