% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-instaloader.R
\name{get_instaloader_version}
\alias{get_instaloader_version}
\title{Check the version of instaloader}
\usage{
get_instaloader_version(envname = NULL)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment. Default
is to use "r-instr" unless set in environment variable instr_PYTHON
(through \code{Sys.setenv()}).}
}
\value{
string with version of Whisper
}
\description{
Check the version of instaloader
}
