% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post.R
\name{get_post}
\alias{get_post}
\title{Get posts by URL or shortcode}
\usage{
get_post(x)
}
\arguments{
\item{x}{URLs or shortcodes}
}
\value{
tibble with metadata about posts
}
\description{
Get posts by URL or shortcode
}
\examples{
\dontrun{
get_post("https://www.instagram.com/p/C5gQOiiC08D/")
}
}
