% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLocation.R
\name{getLocation}
\alias{getLocation}
\title{Get basic information about a location using a location ID}
\usage{
getLocation(locid, token)
}
\arguments{
\item{token}{An OAuth token created with \code{instaOAuth}.}

\item{location}{id.}
}
\description{
\code{getLocation} retrieves location information
}
\examples{
\dontrun{
## See examples for instaOAuth to know how token was created.
## Capturing information about a location
 load("my_oauth")
 loc_id_info <- getLocation( location=423423, token=my_oauth,)
}
}
\author{
Jonne Guyt \email{j.y.guyt@uva.nl}
}

