% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyscale.R
\name{greyscale}
\alias{greyscale}
\title{Greyscale}
\usage{
greyscale(input_path, output_path)
}
\arguments{
\item{input_path}{string, path for the input png file}

\item{output_path}{string, path for the output png file}
}
\value{
a png file at the specified output path
}
\description{
Convert a color image into greyscale
}
\examples{
greyscale("img.png", "img_gs.png")
}
