% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-instaloader.R
\name{init_instr}
\alias{init_instr}
\title{init_instr}
\usage{
init_instr(envname = NULL)
}
\arguments{
\item{envname}{initialise in a specific virtual environment}
}
\description{
Initializes the instaloader python functions. Only works when \code{install_instr()} has ben executed.
}
