% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{insta_login}
\alias{insta_login}
\title{Log into Instagram}
\usage{
insta_login(user, password, save = TRUE, force_refresh = FALSE)
}
\arguments{
\item{user, password}{Specify your Instagram user and password. It's better to
leave this empty, in which case a secure prompt will ask you for input.}

\item{save}{save your session to the cache, so you don't need to specify
password again.}

\item{force_refresh}{overwrite existing sessions in cache.}
}
\description{
Log into Instagram
}
