% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-instaloader.R
\name{install_instaloader}
\alias{install_instaloader}
\title{install instaloader}
\usage{
install_instaloader(envname = NULL, instaloader_version = NULL, ask = TRUE)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment. Default
is to use "r-instr" unless set in environment variable instr_PYTHON
(through \code{Sys.setenv()}).}

\item{instaloader_version}{give a specific version if you want.}

\item{ask}{set \code{FALSE} for unattended install.}
}
\value{
\code{TRUE} if everything worked.
}
\description{
Installs the instaloader Python module
}
