% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_instar}
\alias{install_instar}
\title{install_instar}
\usage{
install_instar(envname = NULL)
}
\arguments{
\item{envname}{specify Python environment name for module installation}
}
\description{
Installs the instaloader Python module
}
