% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url2shortcode}
\alias{url2shortcode}
\title{Convert an Instagram URL to a shortcode}
\usage{
url2shortcode(url)
}
\arguments{
\item{url}{vector of URLs}
}
\value{
character vector of shortcodes
}
\description{
Convert an Instagram URL to a shortcode
}
\examples{
url2shortcode("https://www.instagram.com/p/C5gQOiiC08D/")
}
