% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip}
\usage{
flip(input_path, direction, output_path)
}
\arguments{
\item{input_path}{string, path for the input png file}

\item{direction}{string, direction of flip}

\item{output_path}{string, path for the output png file}
}
\value{
a png file at the specified output path
}
\description{
Flip
}
\examples{
#' flip("img.png", "h", "flipped.png")
}
