% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocation.R
\name{getLocation}
\alias{getLocation}
\title{Get basic information about a location using a location ID}
\usage{
getLocation(location_id, token)
}
\arguments{
\item{location_id}{numeric, location id.}

\item{token}{An OAuth token created with \code{instaOAuth}.}
}
\description{
\code{getLocation} retrieves location information
}
\examples{
\dontrun{
## See examples for instaOAuth to know how token was created.
## Capturing information about a location
 load("my_oauth")
 loc_id_info <- getLocation( location_id=423423, token=my_oauth)
}

}
\author{
Jonne Guyt \email{j.y.guyt@uva.nl}
}

