% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTagCount.R
\name{getTagCount}
\alias{getTagCount}
\title{Get number of times a tag has been used}
\usage{
getTagCount(tag, token)
}
\arguments{
\item{tag}{Hashtag used to filter media.}

\item{token}{An OAuth token created with \code{instaOAuth}.}
}
\description{
\code{getTagCount} retrieves a count of the number of times a tag has
been used in comments to a picture or video.
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

