% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUser.R
\name{getUser}
\alias{getUser}
\title{Get basic information about a user}
\usage{
getUser(username, token, userid = NULL)
}
\arguments{
\item{username}{String, screen name of user.}

\item{token}{An OAuth token created with \code{instaOAuth}.}

\item{userid}{Numeric ID of user.}
}
\description{
\code{getUser} retrieves public information about Instagram user
}
\details{
IMPORTANT: After June 1st, 2016 only applications that have passed permission
review by Instagram will be allowed to access data for users other than the
authenticated user. See  \url{https://www.instagram.com/developer/review/} 
for more information.
}
\examples{
\dontrun{
## See examples for instaOAuth to know how token was created.
## Capturing information about @barackobama
 load("my_oauth")
 obama <- getUser( username="barackobama", token=my_oauth,)
}

}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

