% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyscale.R
\name{greyscale}
\alias{greyscale}
\title{Greyscale}
\usage{
greyscale(img)
}
\arguments{
\item{img}{string of path for an image file in .jpg, .jpeg, .png, .tiff format}
}
\value{
an image file in .jpg, .jpeg, .png, .tiff format
}
\description{
Convert a color image into greyscale
}
\examples{
greyscale("./img.jpg")
}
