% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-instaloader.R
\name{init_instar}
\alias{init_instar}
\title{init_instar}
\usage{
init_instar(envname = NULL)
}
\arguments{
\item{envname}{initialise in a specific virtual environment}
}
\description{
Initializes the instaloader python functions. Only works when \code{install_instar()} has ben executed.
}
